requirements.txt
Django==2.0.7
image==1.5.24
Pillow==5.2.0
pytz==2018.5